back/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==43.0.3
django==5.2.5
django-allauth==65.10.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-nested-routers==0.94.2
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
httplib2==0.22.0
idna==3.10
kombu==5.5.4
narwhals==2.0.1
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.2.0
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pyjwt==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
qrcode==8.2
redis==6.4.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
stripe==12.4.0
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13